Fixed beatRepeat bug (int recordTime vs Double startRepeatTime gave differences in time of note at the start of the loop. Changed recordTime to Double).  Pressing repeat while holding padButton stops note and cancels future noteOff (created dePressedNotesOnRepeat array to track holded pads to cancel).  Fixed playingNotes == -1 when releasing dragged note that was playing using same array.

+++ app/src/main/java/com/example/mysequencer01ui/Classes.kt
@@ -1,7 +1,7 @@
 package com.example.mysequencer01ui
 
 data class Note (
-    var time: Int,
+    var time: Double,
     var pitch: Int,
     var velocity: Int,
     var id: Int,
